requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# API Clients
httpx==0.26.0
apify-client==1.6.4

# Transcription
faster-whisper==0.10.0
yt-dlp==2024.1.29

# Google Sheets
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gspread==6.0.2

# Background Jobs
celery==5.3.6
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
